✨ Feat: simplify gcal parsing using categorizeEvents util
